feat: Actualizar función de eliminación de usuario en el controlador de usuarios
Se ha actualizado la función de eliminación de usuario en el controlador de usuarios para incluir la verificación de autorización utilizando el token JWT decodificado. Esto garantiza que solo los usuarios autorizados puedan eliminar usuarios. Además, se ha modificado el parámetro de entrada para incluir el correo electrónico del usuario a eliminar.

+++ paginaWeb/back/rest/models/users.js
@@ -66,7 +66,6 @@ export class UserModel {
   }
 
   static getAllUsers = async ({ userEmail }) => {
-    // preguntar a dani si meto esto en una función y la llamo desde el controlador :D
     const role = await prisma.users.findUnique({
       where: {
         email: userEmail
@@ -309,8 +308,19 @@ export class UserModel {
     }
   }
 
-  static updateClient = async ({ input }) => {
+  static updateClient = async ({ userEmail, input }) => {
     try {
+      const role = await prisma.users.findUnique({
+        where: {
+          email: userEmail
+        },
+        select: {
+          role: true
+        }
+      })
+      if (role.role.toLowerCase() !== 'a') {
+        throw new Error('Unauthorized')
+      }
       const { email, validated } = input
       await prisma.clients.update({
         where: {
@@ -325,8 +335,20 @@ export class UserModel {
     }
   }
 
-  static deleteUser = async ({ email }) => {
+  static deleteUser = async ({ userEmail, email }) => {
     try {
+      const role = await prisma.users.findUnique({
+        where: {
+          email: userEmail
+        },
+        select: {
+          role: true
+        }
+      })
+      if (role.role.toLowerCase() !== 'a') {
+        throw new Error('Unauthorized')
+      }
+
       await prisma.users.delete({
         where: {
           email
